Added something resembling a cmdline interface to the new vorbiscomment example code. You can say
   vorbiscomment [-l] <file>

and have it dump the comments to stdout, and

   vorbiscomment -w <infile> <outfile>

and have it read a replacement set of comments from stdin.

Still needs some polishing, but it's at least marginally useful
now. I also tried to lay it out so it remained easy to read as
a code example.

The Makefile has been changed to build/install vorbiscomment by
default again, mostly on Monty's approval.