Fix the conditional build stuff.
First, do configure substitution on the top-level subdirs, with DIST_SUBDIRS backup to include everything in 'make dist'. This is cleaner than the previous method of conditionalizing the program-specific makefiles.

Second, switch to configure addition of conditional obj files for optional codec support in ogg123. the AM_CONDITIONAL source method is cleaner for this since the build details are all in the Makefile. However, this only really works in automake 1.7; earlier versions require the matching object files to be in the _LDADD and _DEPENDENCIES lines. If configure is generating these anyway, we might as well simplify the Makefile.am's to rely entirely on substitution.

Also, check for the OggFLAC library as well as FLAC, since these can be installed independently on some distributions.