Require automake 1.6.
nodist_include_HEADERS is not supported by automake 1.4. It ignores the
line entirely and fails to install ao/os_types.h, preventing any
application from compiling against libao. Therefore we make the
requirement explicit by adding it to the AUTOMAKE_OPTIONS in
Makefile.am.

Unfortunately, automake 1.4 is still the default on many systems
(including Debian unstable) and in stark contrast to the situation with
autoconf, even when 1.6 or later is installed there is no dispatch
mechanism to ensure the appropriate version is used. Therefore we make
an attempt in autogen.sh to select the correct executables for aclocal
and automake and run those.

Note that while the message says, "checking for automake 1.6 or
later..." it actually only checks for a matching version number. If
someone knows how to do the 'or later' part in portable shell, please
let me know.