Link png2theora with -lm for rint() in the scons build.
This isn't required on x86_64, probably because gcc converts
it into an intrinsic, so this wasn't noticed until I tried
to build on 32 bit, where gcc 4.4.4 generated an explicit
call to lrint(). In any case, the manpage clearly indicates
linking with -lm, as in fact we do in the autotools build.